Feat/supply

+++ Coffee.QR-BackEnd/Coffee.QR.Core/Domain/OrderItem.cs
@@ -9,14 +9,14 @@ namespace Coffee.QR.Core.Domain
 {
     public class OrderItem : Entity
     {
-        public double Quantity { get; private set; }
+        public long Quantity { get; private set; }
         public long OrderId { get; private set; }
         public Order OrderOrigin { get; private set; }
         public long ItemId { get; private set; }
         public Item ItemPicked { get; private set; }
 
 
-        public OrderItem(double quantity,long orderId, long itemId)
+        public OrderItem(long quantity,long orderId, long itemId)
         {
             Quantity = quantity;
             OrderId = orderId;
